Move `std::io::Cursor` into `core::io`
